requirements.txt
kivy>=2.2.0
PyGithub>=2.1.0
GitPython>=3.1.30
python-dotenv>=1.0.0
cryptography>=41.0.0
requests>=2.31.0
openai>=1.7.0
